requirements.txt
Flask==2.0.2
flask-restx==0.5.1
flask_jwt_extended==4.2.3
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
psycopg2-binary==2.9.2